Move functions defined in main.rs to new files.rs
Creates a new module files.rs to define some functions that were defined
in main.rs. These functions could be useful to someone trying to make a
frontend for Databind, and there's not much reason to keep them in
main.rs.
